Add integer exponential subparser
This change injects a subparser between the layer for
multiplcation/division operators and the next lowest precedence level
(not, unary minus, etc.). As we develop the logic for floats and float
exponentials, it will probably make more sense for this parser to be the
parent parser with two subparsers for float and int exponentials,
respectively.