Add support for List Indexing
This change adds the list indexing operator to our grammar and the
evaluation logic. Also, this change adds a helper method `validListIndex`
which abstracts away the error handling for "index out of bounds"
errors. This works well for now, but we should consider implementing a
more generalized error handler as we start working on the parser error
messages task.